consolidate shared test classes. fix references.
