+++ Src/DependencyCollector/Shared.Tests/Implementation/TelemetryDiagnosticSourceListenerTests.cs
@@ -8,10 +8,11 @@
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.DependencyCollector.Implementation;
     using Microsoft.ApplicationInsights.Extensibility;
-#if !NETCORE
     using Microsoft.ApplicationInsights.Web.TestFramework;
+#if !NETCORE
+    
 #else
-    using Microsoft.ApplicationInsights.Tests;
+    //using Microsoft.ApplicationInsights.Tests;
 #endif
     using Microsoft.VisualStudio.TestTools.UnitTesting;
