+++ Src/DependencyCollector/Shared/DependencyCollector.Shared.projitems
@@ -40,10 +40,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)Implementation\SdkVersionUtils.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Implementation\EventHandlers\EventHubsDiagnosticsEventHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Implementation\EventHandlers\ServiceBusDiagnosticsEventHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Implementation\SqlClientDiagnostics\SqlClientDiagnosticFetcherTypes.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Implementation\SqlClientDiagnostics\SqlClientDiagnosticSourceListener.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Properties\AssemblyInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PropertyFetcher.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SanitizedHostList.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)SqlClientDiagnosticSourceListener.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TelemetryDiagnosticSourceListener.cs" />
   </ItemGroup>
   <ItemGroup Condition=" '$(TargetFramework)' != 'netstandard1.6' ">
