+++ Src/PerformanceCollector/Perf.NetFull.Tests/Perf.NetFull.Tests.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), 'Test.props'))\Test.props" />
   <PropertyGroup>
@@ -6,7 +6,6 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{F254D4FB-428D-408E-8251-39BCA7B4B5CE}</ProjectGuid>
     <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Microsoft.ApplicationInsights.Tests</RootNamespace>
     <AssemblyName>Microsoft.AI.PerformanceCollector.NetFull.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
@@ -86,8 +85,6 @@
     <Compile Include="PerformanceCollectorModuleTests.cs" />
     <Compile Include="PerformanceCollectorEventSourceTests.cs" />
     <Compile Include="StandardPerformanceCollectorTests.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Mocks\StubTelemetryChannel.cs" />
     <Compile Include="TelemetryAction.cs" />
     <Compile Include="WebAppPerformanceCollector\AzureWebAppTest.cs" />
     <Compile Include="WebAppPerformanceCollector\CacheHelperTests.cs" />
@@ -118,7 +115,6 @@
   <ItemGroup>
     <EmbeddedResource Include="QuickPulse\PerfLib\PerfData.data" />
   </ItemGroup>
-  <ItemGroup />
   <Import Project="..\Perf.Shared.Tests\Perf.Shared.Tests.projitems" Label="Shared" />
   <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
@@ -130,4 +126,5 @@
   </Target>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), 'Common.targets'))\Common.targets" />
   <Import Project="..\..\..\..\packages\StyleCop.MSBuild.4.7.54.0\build\StyleCop.MSBuild.Targets" Condition="Exists('..\..\..\..\packages\StyleCop.MSBuild.4.7.54.0\build\StyleCop.MSBuild.Targets')" />
+  <Import Project="..\..\TestFramework\Shared\TestFramework.Shared.projitems" Label="Shared" />
 </Project>
