--- Src/PerformanceCollector/Perf.NetFull.Tests/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyCompany("Microsoft")]
-[assembly: AssemblyCopyright("Copyright © Microsoft. All Rights Reserved.")]
-
-[assembly: ComVisible(false)]
