+++ Src/TestFramework/Shared/StubTransmission.cs
@@ -6,9 +6,9 @@
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
 
-#if NET45
+//#if NET45
     using TaskEx = System.Threading.Tasks.Task;
-#endif
+//#endif
 
     internal class StubTransmission : Transmission
     {
